rename Template to Fill, move to core, add a bit of docstring, remove dead TODO in core
