+++ glom/__init__.py
@@ -3,6 +3,7 @@ from glom.core import (glom,
                        register,
                        Glommer,
                        Call,
+                       Fill,
                        Spec,
                        OMIT,  # backwards compat
                        SKIP,
@@ -22,7 +23,6 @@ from glom.core import (glom,
 
 from glom.reduction import Sum, Fold, Flatten, flatten, FoldError, Merge, merge
 from glom.mutation import Assign, assign, PathAssignError
-from glom.template import Template
 
 # there's no -ion word that really fits what "streaming" means.
 # generation, production, iteration, all have more relevant meanings
